refactor(extension): split sidepanel control components
